[HAR-83] peer discovery works under account model
fix a bug to update the bitmap on the leader in account model

Issue:

Can only reach consensus once, and then got the following errors.

Failed to create new block               Error="nonce too low"

This is because of new validators joined the shard and their state are
not sync'ed, thus we have not enough nonce.

Signed-off-by: Leo Chen <leo@harmony.one>