Fetch and use (“gimme”) stable Go, not master
The Go installation (gimme) step of Travis-CI is super slow (~3m),
because we use master version of Go, which gimme needs to
download/compile/install out of the Git repository.

Use stable instead of master, which lets gimme to download/install the
binary version (if available), of the latest stable release, and fall
back to source only if that fails.