Tag libp2p packages with libp2p_ prefix
This is because we are introducing more and more parallel P2P concepts
(libp2p peer versus our own peer, for example).  Explicit tagging makes
it easier for untrained eyes to understand which side the code refers
to: If it's prefixed with libp2p_, it's libp2p; if not, it's our p2p.

While we are at this, apply the std → 3rd-party → local import grouping,
and also rename multiaddr to ma (the upstream canonical name of
go-multiaddr) in the affected files.