Fix key issues and remove storing triggered state
I removed triggered state because how it was functioning before would cause potentially some conflicts with manual pagerduty resolutions. For example, if a node broke, it would save true into triggered, and then only trigger again if it was set to false. It could only be set to false if the node came back up, which, when the script ran again, false would be saved into triggered. However, if the alert was accidentally resolved and the node was still broken, no alert would come up as triggered would still have true. The save state was mainly to prevent multiple posts, but as there is the dedup key and the program doesn't stall if there is no alert to resolve, there's no real reason to have it.