Enable cross-shard transaction at CrossTxEpoch+1
This is to allow for all 4 shards to roll into CrossTxEpoch and start
having cross-shard-transaction-related fields in block header/body
before the fields are actually used.  Otherwise, if a shard A enters
CrossTxEpoch before shard B does and sends a cross-shard receipt to
shard B, it cannot be processed because chain B still lacks the incoming
receipts field.