[libp2p/consensus] p2p spamming defense
This is a big PR merged many small commits together.
We add the message validation function in libp2p layer.

In the validation function, we check the following conditions
1) the p2p message is a valid consensus message
2) the p2p message sender has a valid public key
3) the sender's public key is in the current committee
4) log the number of invalid/valid messages

After the validation, the valid messages will be forward to the network,
while the invalid messages will be filtered out.

The messages intended for the validator will be handled in the consensus layer.

Signed-off-by: Leo Chen <leo@harmony.one>