feat: Prune stale staking data hard fork + onwards
Stale staking data is defined as delegations to a validator which have
no rewards, no amount, and no undelegations. This definition, however,
excludes the (inactive) validator's self delegation since it is expected
to be at index 0, and a validator cannot be deleted at the moment. The
process works as below:

 - At the last block of every epoch including the hard fork epoch,
   obtain a list of validators. For each validator, load the
   ValidatorWrapper and iterate through Delegations to find stale
   delegations. Drop such delegations, and make a map of delegators as
   keys with validator addres(es) as the values.
 - Using the map obtained in (1), clear the delegations by delegator in
   offchain data. The map is therefore cached in the ProcessorResult for
   ease of access (between state_processor and node_newblock).
 - Since the objective of this change is to save some space following
   the hard fork, do not allow small undelegations. This means that
   undelegations must be made in such a manner that the remaining amount
   after the undelegation has been made is at least a 100 ONEs, or zero.
 - Lastly, since the removal results in different lengths between
   delegations of validator snapshots and validator wrappers, modify the
   algorithm for AddReward to be based on DelegatorAddress rather than
   the loop index. See core/state/statedb.go

The impact of this change on block processing speed is small. I ran a
node with this change at epoch 881 on mainnet using the rclone database
and saw that pruneStaleStakingData takes 18ms to complete (if each prune
is not logged individually, it takes 9ms).

The PR has been tested locally for integration testing (the undelegation
bit), and unit tests for pruning stale data, remaining undelegation, and
AddReward have been added. Goimports is included.