feat: Prune validator rewards in archival node...
...across all blocks. Whenever a new block is saved to state, go an
epoch behind to iterate over all validators during that block and clear
the Code field.

In other words, existing validator wrappers on disk, from one epoch ago,
are replaced with nil code. This is activated either through a manual
flag (`--prune.rewards`) for nodes which are already in sync, or auto
activated for other archival nodes when merged.