Increment version on dependency, to fix build on go 1.18
The version of the reflect2 library we were pulling in (transitively,
via json-iterator) segfaults on go 1.18, due to some use of the unsafe
package that is not portable accross version; I was seeing this manifest
not long after running `make debug`.

The problem is fixed in the latest version, and similarly the latest
version of json-iterator depends on the fixed version. So, this patch
updates to the latest version to  get he fix.