Fix the decode test.
The problem was that we were doing args := input[4:] to chop off the
method id, but then continuing to read from *input* everywhere, instead
of args -- so everything was shifted 4 bytes. It appears to work now,
but we should test some other exmaples.

This also remove some stray debug prints, and adds some utility code to
the tests to format the payload nicely, which was helpful when figuring
all this out.