[core] fix: cross shard precompile gas to 0...
...and ensure transfers from the precompile address can be made before
executing them. Since we are now blocking multiple `CXReceipts` per
transaction explicitly, it doesn't make sense to be wary of contracts
subsidizing cross shard transfers for end users via `delegatecall`. It
is then safe to remove the gas charge for the precompile. Secondly,
ensure that there are enough funds on the precompile via `CanTransfer`
as part of the sanity check.