[config] change default localnet sync config...
...to allow localnet cross shard tests to pass. It has been observed
that the epoch block from shard 0, which contains the `ShardState` does
not reach shard 1 on localnet in time. This leads to the cross shard
receipt not being validated, and thus, the block is not signed by one
node (8.5% vote power) in some cases. This change makes the sync
parameters a bit aggressive (only on localnet) to reduce the number of
times the test can fail.