Add logging for `NthNextHmy` panic (#4341)
* improve logging

* [consensus] do not try to use negative index

* Revert "do not try to use negative index"

This reverts commit b434fd3f4af39f32650e909292cc0123bedba86e. We have
fixed the cause of the issue, which was time drift on a new cloud
provider's nodes. See `systemd-timesyncd.service`

Even if this fix had been merged, it would likely not have solved the
problem given those nodes with the correct time would pick a different
leader from those with time drift. Or, in other words, the view change
would not have gone through.

* improve logging